the_best

+++ dav/views.py
@@ -4,9 +4,10 @@ from django.shortcuts import get_object_or_404
 from rest_framework import status 
 from .serializers import * 
 from .models import * 
-from rest_framework.decorators import api_view 
+from rest_framework.decorators import api_view,parser_classes
 from minio import Minio
 from datetime import datetime
+from django.http import HttpResponseBadRequest,HttpResponseServerError
 from rest_framework import viewsets
 from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
@@ -19,6 +20,7 @@ from django.contrib.auth import get_user_model
 import redis
 import uuid
 from django.conf import settings
+from rest_framework.parsers import MultiPartParser
 
 
 session_storage = redis.StrictRedis(host=settings.REDIS_HOST, port=settings.REDIS_PORT)
@@ -42,6 +44,31 @@ def create(request):
         return Response({'status': 'Success'}, status=200)
     return Response({'status': 'Error', 'error': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
+@api_view(['GET'])
+@permission_classes([IsAuth])
+def user_info(request):
+    print(request.headers.get('Authorization'))
+    try:
+        access_token = request.COOKIES["access_token"]
+        print('access_token', access_token)
+        if session_storage.exists(access_token):
+            email = session_storage.get(access_token).decode('utf-8')
+            user = Users.objects.get(email=email)
+            application = Applications.objects.filter(customer_id=user.id).filter(status=1).first()
+            user_data = {
+                "user_id": user.id,
+                "email": user.email,
+                "is_moderator": user.is_moderator,
+                "current_cart": application.id if application else -1,
+            }
+            print(user_data)
+            return Response(user_data, status=status.HTTP_200_OK)
+        else:
+            return Response({'status': 'Error', 'message': 'Session does not exist'})
+    except:
+        return Response({'status': 'Error', 'message': 'Cookies are not transmitted'})
+
+
 
 # @authentication_classes([])
 # @csrf_exempt
@@ -55,29 +82,34 @@ def login_view(request):
     
     if user is not None:
         random_key = str(uuid.uuid4())
+        application = Applications.objects.filter(customer_id=user.id).filter(status=1).first()
         user_data = {
             "user_id": user.id,
             "email": user.email,
             "is_moderator": user.is_moderator,
-            "session_id": random_key
+            "access_token": random_key,
+            "current_cart": application.id if application else -1,
         }
         session_storage.set(random_key, username)
         response = Response(user_data, status=status.HTTP_201_CREATED)
-        response.set_cookie("session_id", random_key)
+        response.set_cookie("access_token", random_key)
 
         return response
     else:
         return Response({'status': 'Error'}, status=status.HTTP_400_BAD_REQUEST)
 
+@api_view(['Post'])
+@permission_classes([IsAuth])
 def logout_view(request):
-    authorization_header = request.headers.get('Authorization')
-    access_token = authorization_header.split(' ')[1] if authorization_header else None
+    # authorization_header = request.headers.get('Authorization')
+    # access_token = authorization_header.split(' ')[1] if authorization_header else None
+    access_token = request.COOKIES["access_token"]
     if access_token is None:
         message = {"message": "Token is not found in cookie"}
         return Response(message, status=status.HTTP_401_UNAUTHORIZED)
     session_storage.delete(access_token)
     response = Response({'message': 'Logged out successfully'})
-    response.delete_cookie('session_id')
+    response.delete_cookie('access_token')
 
     return response
 
@@ -87,8 +119,10 @@ def logout_view(request):
 
 
 
+
 #GET - получить список всех опций 
 @api_view(['Get']) 
+@permission_classes([AllowAny])
 def get_options(request, format=None): 
     search_query = request.GET.get('search', '')
     faculty = request.GET.get('faculty', '')
@@ -98,25 +132,28 @@ def get_options(request, format=None):
     if faculty and faculty != 'Любой факультет':
         options = options.filter(faculty=faculty)
     
-    serializer = OptionSerializer(options, many=True)
-    
-    #Retrieve the application with customer user and status equal to 1
-    application = Applications.objects.filter(customer=user, status=1).first()
-    if application:
-        application_serializer = ApplicationSerializer(application)
-        apps_data = [application_serializer.data]
-    else:
-        apps_data = []
-    
-    response_data = {
-        'apps': apps_data,
-        'options': serializer.data,
-    }
-    
-    return Response(response_data)
+    try:
+        access_token = request.COOKIES["access_token"]
+        username = session_storage.get(access_token).decode('utf-8')
+        print(username)
+        user_ind = Users.objects.filter(email=username).first()
+        application = Applications.objects.filter(customer_id=user_ind.id, status=1).values_list('id', flat=True).first()
+        serializer = OptionSerializer(options, many=True)
+        response_data = {
+            'app_id': application,
+            'options': serializer.data,
+        }
+        return Response(response_data)
+    except:
+        serializer = OptionSerializer(options, many=True)
+        result = {
+            'options': serializer.data,
+        }
+        return Response(result)
 
 #POST - добавить новую опцию  
-@api_view(['Post']) 
+@api_view(['Post'])
+@permission_classes([IsModerator])
 def post_option(request, format=None):     
     serializer = OptionSerializer(data=request.data) 
     if not serializer.is_valid(): 
@@ -145,9 +182,43 @@ def post_option(request, format=None):
     serializer = OptionSerializer(option, many=True)
     return Response(serializer.data)
 
+@api_view(['POST'])
+@parser_classes([MultiPartParser])
+@permission_classes([AllowAny])
+def postImageToOption(request, pk):
+    if 'file' in request.FILES:
+        file = request.FILES['file']
+        subscription = Options.objects.get(pk=pk, available=True)
+        
+        client = Minio(endpoint="localhost:9000",
+                       access_key='minioadmin',
+                       secret_key='minioadmin',
+                       secure=False)
+
+        bucket_name = 'img'
+        file_name = file.name
+        file_path = "http://localhost:9000/img/" + file_name
+        
+        try:
+            client.put_object(bucket_name, file_name, file, length=file.size, content_type=file.content_type)
+            print("Файл успешно загружен в Minio.")
+            
+            serializer = OptionSerializer(instance=subscription, data={'image': file_path}, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return HttpResponse('Image uploaded successfully.')
+            else:
+                return HttpResponseBadRequest('Invalid data.')
+        except Exception as e:
+            print("Ошибка при загрузке файла в Minio:", str(e))
+            return HttpResponseServerError('An error occurred during file upload.')
+
+    return HttpResponseBadRequest('Invalid request.')
+
 
 #GET - получить одну опцию 
-@api_view(['Get']) 
+@api_view(['Get'])
+@permission_classes([AllowAny]) 
 def get_option(request, pk, format=None): 
     option = get_object_or_404(Options, pk=pk) 
     if request.method == 'GET': 
@@ -156,6 +227,7 @@ def get_option(request, pk, format=None):
  
 #PUT - обновить одну опцию 
 @api_view(['Put']) 
+@permission_classes([AllowAny])
 def put_option(request, pk, format=None): 
     option = get_object_or_404(Options, pk=pk) 
     serializer = OptionSerializer(option, data=request.data) 
@@ -166,6 +238,7 @@ def put_option(request, pk, format=None):
  
 #PUT - удалить одну опцию 
 @api_view(['Put']) 
+@permission_classes([IsModerator])
 def delete_option(request, pk, format=None):     
     if not Options.objects.filter(pk=pk).exists():
         return Response(f"Опции с таким id не существует!") 
@@ -179,7 +252,14 @@ def delete_option(request, pk, format=None):
  
 #POST - добавить услугу в заявку(если нет открытых заявок, то создать)
 @api_view(['POST'])
+@permission_classes([IsAuth])
 def add_to_application(request, pk):
+    access_token = request.COOKIES["access_token"]
+    username = session_storage.get(access_token).decode('utf-8')
+    user = Users.objects.filter(email=username).first()
+    if user is None:
+        print('Не зарегестрирован')
+
     if not Options.objects.filter(id=pk).exists():
         return Response(f"Услуги с таким id не существует!")
 
@@ -228,14 +308,29 @@ def add_to_application(request, pk):
  
 #GET - получить список всех заявок 
 @api_view(['Get']) 
+@permission_classes([IsAuth])
 def get_applications(request, format=None): 
-    print('get') 
-    applications = Applications.objects.all() 
-    serializer = ApplicationSerializer(applications, many=True) 
-    return Response(serializer.data) 
+
+    access_token = request.COOKIES["access_token"]
+    username = session_storage.get(access_token).decode('utf-8')
+    user_id = Users.objects.filter(email=username).values_list('id', flat=True).first()
+
+    if username is not None and user_id is not None:
+        user = Users.objects.get(email=username)
+        print(user.is_moderator)
+        if user.is_moderator:
+            applications = Applications.objects
+        else:
+            applications = Applications.objects.filter(customer_id=user_id).exclude(status=2)
+        serializer = ApplicationSerializer(applications, many=True)
+        return Response(serializer.data)
+    else:
+        return Response("Invalid user", status=status.HTTP_400_BAD_REQUEST)
  
+
 #GET - получить одну заявку 
 @api_view(['GET'])
+@permission_classes([IsAuth])
 def get_application(request, pk, format=None):
     application = get_object_or_404(Applications, pk=pk)
     if request.method == 'GET':
@@ -260,6 +355,7 @@ def get_application(request, pk, format=None):
 
 
 @api_view(["PUT"]) 
+@permission_classes([AllowAny])
 def update_by_user(request, pk): 
     if not Applications.objects.filter(pk=pk).exists(): 
         return Response(f"Заявки с таким id не существует!") 
@@ -279,11 +375,21 @@ def update_by_user(request, pk):
     application.status = request_status 
     application.save() 
  
-    serializer = ApplicationSerializer(application, many=False) 
-    return Response(serializer.data)
+    serializer = ApplicationSerializer(application, many=False)
+    response = Response(serializer.data)
 
+    return response
+
+
+@swagger_auto_schema(method='put',request_body=ApplicationSerializer)
 @api_view(["PUT"]) 
+@permission_classes([IsModerator])
 def update_by_admin(request, pk): 
+
+    access_token = request.COOKIES["access_token"]
+    modername = session_storage.get(access_token).decode('utf-8')
+    user = Users.objects.filter(email=modername).first()
+
     if not Applications.objects.filter(pk=pk).exists(): 
         return Response(f"Заявки с таким id не существует!") 
  
@@ -300,15 +406,19 @@ def update_by_admin(request, pk):
     # if app_status == 5: 
     #     return Response("Статус изменить нельзя") 
  
-    application.status = request_status 
-    application.save() 
- 
-    serializer = ApplicationSerializer(application, many=False) 
-    return Response(serializer.data)
+    application.moderator_id=user.id
+    application.status = request_status
+    application.save()
+
+    serializer = ApplicationSerializer(application, many=False)
+    response = Response(serializer.data)
+    response.setHeader("Access-Control-Allow-Methods", "PUT")
+    return response
 
 
 #DELETE - удалить одну заявку
 @api_view(['Delete']) 
+@permission_classes([AllowAny])
 def delete_application(request, pk, format=None):     
     application = get_object_or_404(Applications, pk=pk) 
     print(application.status)
@@ -322,6 +432,7 @@ def delete_application(request, pk, format=None):
 
 #DELETE - удалить конкретную услугу из конкретной заявки
 @api_view(["DELETE"])
+@permission_classes([AllowAny])
 def delete_option_from_application(request, application_id, option_id):
     if not Applications.objects.filter(pk=application_id).exists():
         return Response("Заявки с таким id не существует", status=status.HTTP_404_NOT_FOUND)
@@ -332,24 +443,11 @@ def delete_option_from_application(request, application_id, option_id):
     application = Applications.objects.get(pk=application_id)
     option = Options.objects.get(pk=option_id)
 
+    application_subscription = get_object_or_404(Applicationsoptions, application=application, option=option)
+    if application_subscription is None:
+        print('yeeeeee')
+        return Response("Заявка не найдена", status=404)
     application.applicationsoptions_set.filter(option=option).delete()
     application.save()
 
-    return Response("Опция успешно удалена из заявки", status=status.HTTP_204_NO_CONTENT)
-
-
-# #PUT - изменить кол-во конкретной опции в заявке
-# @api_view(["PUT"])
-# def update_option_amount(request, application_id, option_id):
-#     if not Applications.objects.filter(pk=application_id).exists() or not Options.objects.filter(pk=option_id).exists():
-#         return Response("Заявки или опции с такими id не существует", status=status.HTTP_404_NOT_FOUND)
-
-#     application_option = Applicationsoptions.objects.filter(application_id=application_id, option_id=option_id).first()
-
-#     if not application_option:
-#         return Response("В этой заявке нет такой опции", status=status.HTTP_404_NOT_FOUND)
-
-#     new_amount = request.data.get("amount",1)
-#     application_option.amount = new_amount
-#     application_option.save()
-#     return Response("Amount успешно обновлен", status=status.HTTP_200_OK)
+    return Response("Опция успешно удалена из заявки", status=200)
